refactor: decompose studio.api.ts into 4 modules (Sprint 051)
Split 953 LOC monolith into focused modules:
- studio-tracks.api.ts: track versions, stems, source track
- studio-generation.api.ts: generation tasks, section replacement, mashup
- studio-transcription.api.ts: stem transcriptions, MIDI, Klangio/Replicate
- studio-realtime.api.ts: realtime subscriptions, studio projects

Main studio.api.ts re-exports for backward compatibility.
