+++ src/api/studio/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Studio API — decomposed modules
+ * Sprint 051: split from studio.api.ts (953 LOC → 4 modules)
+ */
+export * from "./studio-tracks.api";
+export * from "./studio-generation.api";
+export * from "./studio-transcription.api";
+export * from "./studio-realtime.api";
